+++ cfgov/v1/models/__init__.py
@@ -5,6 +5,7 @@ from .molecules import *
 from .organisms import *
 from .snippets import *
 from .landing_page import *
+from .ref import *
 
 if settings.DEBUG:
     from .demo import *